# backend/src/domains/group/routes.ts
import type { Application, RequestHandler } from 'express';
import type { Pool } from 'pg';
import { randomUUID } from 'crypto';
import { mediaFile } from '../message/media.js';
export function registerGroupRoutes(app:Application,pool:Pool,broadcast:(type:string,payload:any)=>void,auth:RequestHandler,write:RequestHandler){
app.get('/api/media/:id',auth,async(req,res)=>{const item=await mediaFile(pool,String(req.params.id));if(!item)return res.status(404).json({error:{code:'MEDIA_NOT_FOUND',message:'媒体文件不存在',requestId:req.requestId}});res.type(item.contentType).sendFile(item.path)});
app.get('/api/groups',auth,async(req,res)=>{const rows=(await pool.query('SELECT * FROM groups ORDER BY id DESC')).rows;const groups=await Promise.all(rows.map(async(g:any)=>{const members=(await pool.query('SELECT account_id AS "accountId",platform_user_id AS "platformUserId",role FROM group_members WHERE group_id=$1',[g.id])).rows;return {id:g.id,gatewayGroupId:g.gateway_group_id,status:g.status,creatorAccountId:g.creator_account_id,agentEnabled:g.agent_enabled,autoKickEnabled:g.auto_kick_enabled,members,activeSequenceRunId:(await pool.query("SELECT id FROM sequence_runs WHERE group_id=$1 AND status='running' LIMIT 1",[g.id])).rows[0]?.id||null,activeAgentRunId:(await pool.query("SELECT id FROM agent_runs WHERE group_id=$1 AND status='running' LIMIT 1",[g.id])).rows[0]?.id||null}}));res.json(groups)});
app.post('/api/groups',auth,write,async(req,res)=>{
  const {creatorAccountId,memberAccountIds}=req.body||{};
  if(typeof creatorAccountId!=='string'||!Array.isArray(memberAccountIds)||memberAccountIds.length<1||memberAccountIds.some((id:unknown)=>typeof id!=='string')||new Set([creatorAccountId,...memberAccountIds]).size!==memberAccountIds.length+1)return res.status(400).json({error:{code:'VALIDATION_ERROR',message:'群主与成员账号必须有效且不重复',requestId:req.requestId}});
  const ids=[creatorAccountId,...memberAccountIds];
  const accounts=await pool.query('SELECT id,status FROM accounts WHERE id=ANY($1)',[ids]);
  if(accounts.rows.length!==ids.length||accounts.rows.some(row=>row.status!=='online'))return res.status(422).json({error:{code:'ACCOUNT_NOT_ONLINE',message:'全部账号必须在线',requestId:req.requestId}});
  const job=await pool.query("INSERT INTO jobs(status,kind,payload) VALUES('running','create',$1) RETURNING id",[JSON.stringify({creatorAccountId,memberAccountIds})]);
  return res.status(202).json({jobId:job.rows[0].id});
});
app.get('/api/jobs/:id',auth,async(req,res)=>{const q=await pool.query('select * from jobs where id=$1',[req.params.id]);if(!q.rowCount)return res.status(404).json({error:{code:'NOT_FOUND',message:'资源不存在',requestId:req.requestId}});res.json(q.rows[0])});
app.get('/api/groups/:id',auth,async(req,res)=>{const q=await pool.query('select * from groups where id=$1',[req.params.id]);if(!q.rowCount)return res.status(404).json({error:{code:'NOT_FOUND',message:'资源不存在',requestId:req.requestId}});const g=q.rows[0],m=await pool.query('select account_id as "accountId",platform_user_id as "platformUserId",role from group_members where group_id=$1',[req.params.id]);res.json({id:g.id,gatewayGroupId:g.gateway_group_id,status:g.status,creatorAccountId:g.creator_account_id,agentEnabled:g.agent_enabled,autoKickEnabled:g.auto_kick_enabled,members:m.rows,activeSequenceRunId:(await pool.query("select id from sequence_runs where group_id=$1 and status='running' limit 1",[req.params.id])).rows[0]?.id||null,activeAgentRunId:(await pool.query("select id from agent_runs where group_id=$1 and status='running' limit 1",[req.params.id])).rows[0]?.id||null})});
app.patch('/api/groups/:id',auth,write,async(req,res)=>{const fields=[];const vals:boolean[]=[];for(const k of ['agentEnabled','autoKickEnabled'])if(req.body?.[k]!==undefined){if(typeof req.body[k]!=='boolean')return res.status(400).json({error:{code:'VALIDATION_ERROR',message:`${k} 必须是布尔值`,requestId:req.requestId}});fields.push(`${k==='agentEnabled'?'agent_enabled':'auto_kick_enabled'}=$${vals.length+1}`);vals.push(req.body[k])}if(!fields.length)return res.status(400).json({error:{code:'VALIDATION_ERROR',message:'至少提供一个可更新字段',requestId:req.requestId}});const updated=await pool.query(`update groups set ${fields.join(',')} where id=$${vals.length+1} returning id`,[...vals,req.params.id]);if(!updated.rowCount)return res.status(404).json({error:{code:'GROUP_NOT_FOUND',message:'群不存在',requestId:req.requestId}});return res.json({ok:true})});
app.get('/api/groups/:id/messages',auth,async(req,res)=>{const limit=Math.min(Math.max(Number(req.query.limit||50),1),100);let cursor:any=null;try{if(req.query.before){const raw=Buffer.from(String(req.query.before),'base64url').toString('utf8');cursor=JSON.parse(raw)}}catch{return res.status(400).json({error:{code:'VALIDATION_ERROR',message:'无效的消息游标',requestId:req.requestId}})}const args:any[]=[req.params.id];let condition='';if(cursor){if(!cursor.sentAt||!cursor.id)return res.status(400).json({error:{code:'VALIDATION_ERROR',message:'无效的消息游标',requestId:req.requestId}});args.push(cursor.sentAt,cursor.id);condition='AND (sent_at,id)<($2,$3)'}args.push(limit);const result=await pool.query(`SELECT id,msg_id AS "msgId",client_msg_id AS "clientMsgId",sender_platform_user_id AS "senderPlatformUserId",text,sent_at AS "sentAt",delivery_status AS "deliveryStatus",fail_code AS "failCode",is_own AS "isOwn",media_url AS "mediaUrl",CASE WHEN local_file_path IS NULL THEN NULL ELSE '/api/media/'||id::text END AS "localFilePath" FROM messages WHERE group_id=$1 ${condition} ORDER BY sent_at DESC,id DESC LIMIT $${args.length}`,args);const last=result.rows.at(-1);res.json({items:result.rows,nextCursor:result.rows.length===limit&&last?Buffer.from(JSON.stringify({sentAt:last.sentAt,id:last.id})).toString('base64url'):null})});
app.post('/api/groups/:id/send',auth,write,async(req,res)=>{const {accountId,text}=req.body||{};if(typeof text!=='string'||!text.trim()||typeof accountId!=='string')return res.status(400).json({error:{code:'VALIDATION_ERROR',message:'账号和文本必填',requestId:req.requestId}});const group=await pool.query('SELECT status FROM groups WHERE id=$1',[req.params.id]);if(!group.rowCount)return res.status(404).json({error:{code:'GROUP_NOT_FOUND',message:'群不存在',requestId:req.requestId}});if(group.rows[0].status!=='active')return res.status(409).json({error:{code:'GROUP_UNREACHABLE',message:'群不可写',requestId:req.requestId}});const member=await pool.query('SELECT m.platform_user_id,a.status FROM group_members m JOIN accounts a ON a.id=m.account_id WHERE m.group_id=$1 AND m.account_id=$2',[req.params.id,accountId]);if(!member.rowCount)return res.status(409).json({error:{code:'ACCOUNT_NOT_IN_GROUP',message:'账号不在群内',requestId:req.requestId}});if(!['online','rate_limited'].includes(member.rows[0].status))return res.status(409).json({error:{code:'ACCOUNT_UNAVAILABLE',message:'账号不可用',requestId:req.requestId}});const clientMsgId=randomUUID();await pool.query("INSERT INTO messages(group_id,client_msg_id,text,sender_platform_user_id,sent_at,delivery_status,is_own,outbound_account_id) VALUES($1,$2,$3,$4,now(),'queued',true,$5)",[req.params.id,clientMsgId,text,member.rows[0].platform_user_id,accountId]);res.status(202).json({clientMsgId})});
app.post('/api/groups/:id/leave-all',auth,write,async(req,res)=>{
  const group=await pool.query('SELECT id FROM groups WHERE id=$1',[req.params.id]);
  if(!group.rowCount)return res.status(404).json({error:{code:'GROUP_NOT_FOUND',message:'群不存在',requestId:req.requestId}});
  const job=await pool.query("INSERT INTO jobs(status,kind,group_id) VALUES('running','leave_all',$1) RETURNING id",[req.params.id]);
  return res.status(202).json({jobId:job.rows[0].id});
});
function agentRunView(run:any){return {id:run.id,groupId:run.group_id,status:run.status,endReason:run.end_reason,summary:run.summary,steps:run.steps,createdAt:run.created_at}}app.get('/api/agent-runs/:id',auth,async(req,res)=>{const q=await pool.query('SELECT * FROM agent_runs WHERE id=$1',[req.params.id]);if(!q.rowCount)return res.status(404).json({error:{code:'NOT_FOUND',message:'资源不存在',requestId:req.requestId}});res.json(agentRunView(q.rows[0]))});app.get('/api/groups/:id/agent-runs',auth,async(req,res)=>res.json((await pool.query('SELECT * FROM agent_runs WHERE group_id=$1 ORDER BY created_at DESC LIMIT 20',[req.params.id])).rows.map(agentRunView)));
}